requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
isort==4.3.20
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
numpy==1.16.4
opencv-contrib-python==4.1.0.25
Pillow==6.0.0
protobuf==3.8.0
pylint==2.3.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
